Split search into abstract class
