+++ classes/files.php
@@ -23,7 +23,7 @@ namespace tool_advancedreplace;
  * @copyright  2024 Catalyst IT Australia Pty Ltd
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class files extends \core\persistent {
+class files extends search {
 
     /** The name of the database table. */
     public const TABLE = 'tool_advancedreplace_files';
@@ -40,6 +40,12 @@ class files extends \core\persistent {
         'skipfilenames',
         'skipareas',
     ];
+
+    /** @var string File area for output files */
+    protected $filearea = 'files';
+
+    /** @var string Class for the adhoc task */
+    protected $adhoctask = \tool_advancedreplace\task\files::class;
 
     /**
      * Return the definition of the properties of this model.
@@ -107,82 +113,5 @@ class files extends \core\persistent {
                 'default' => 0,
             ],
         ];
-    }
-
-    /**
-     * Hook to execute before a delete.
-     *
-     * @return void
-     */
-    protected function before_delete(): void {
-        // TODO: Clean up any remaining adhoc tasks.
-    }
-
-    /**
-     * Hook to execute after a delete
-     *
-     * @param bool $result Whether or not the delete was successful.
-     * @return void
-     */
-    protected function after_delete($result): void {
-        if ($result) {
-            // Delete output pluginfiles.
-            if ($file = $this->get_file(self::to_record())) {
-                $file->delete();
-            }
-        }
-    }
-
-    /**
-     * Queues a search task to be run
-     * @return bool true if the task was queued
-     */
-    public function queue_task(): bool {
-        $adhoctask = new \tool_advancedreplace\task\files();
-        $adhoctask->set_custom_data([
-            'searchid' => $this->get('id'),
-        ]);
-        return \core\task\manager::queue_adhoc_task($adhoctask);
-    }
-
-    /**
-     * Returns a clean copy of data that can be used to rerun a search.
-     * @return \stdClass
-     */
-    public function copy_data(): \stdClass {
-        $data = new \stdClass();
-        foreach (self::COPY_COLUMNS as $column) {
-            $data->$column = $this->get($column);
-        }
-        return $data;
-    }
-
-    /**
-     * Gets the file for the search output
-     * @param \stdClass $record
-     * @return bool|\stored_file
-     */
-    public static function get_file(\stdClass $record) {
-        $filename = self::get_filename($record);
-        $fs = get_file_storage();
-        return $fs->get_file(
-            \context_system::instance()->id,
-            'tool_advancedreplace',
-            'files',
-            $record->id,
-            '/',
-            $filename
-        );
-    }
-
-    /**
-     * Gets the file name of the search output
-     * @param \stdClass $record
-     * @return string filename
-     */
-    public static function get_filename(\stdClass $record): string {
-        // The hardcoded default filename should not be changed.
-        $name = !empty($record->name) ? $record->name : 'searchresult-' . $record->id;
-        return strtolower($name) . '.csv';
     }
 }
